requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
boto3==1.34.131
botocore==1.34.131
build==1.2.1
CacheControl==0.14.0
certifi==2024.6.2
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.8
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.1.2
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-pagination==0.12.25
fastjsonschema==2.20.0
filelock==3.15.3
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.2.0
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jmespath==1.0.1
keyring==24.3.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
msgpack==1.0.8
orjson==3.10.5
packaging==24.1
pathlib==1.0.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
Pygments==2.18.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.9.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tomlkit==0.12.5
trove-classifiers==2024.5.22
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
virtualenv==20.26.2
watchfiles==0.22.0
websockets==12.0
zipp==3.19.2
